ref: seperating and moving components into different files and folder
